requirements.txt
asgiref==3.8.1
Django==5.1.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
